+++ docs/md/releasenotes.md
@@ -3,6 +3,7 @@
 ### Improved
 
 - Enable new `PartialEvaluator` error message format by default.
+- Add an option for `ClassReferenceFixer` to only rename fields when there is a name clash.
 
 ## Version 9.1